report/requirements.txt
dash-bootstrap-components==1.0.3
dash-daq==0.5.0
dash==2.3.1
dash-leaflet==0.1.13
gunicorn==19.9.0
pandas==1.2.4
psycopg2-binary==2.7.7
pytest==6.2.3
pytz==2021.1
xlrd==1.1.0
easy-thumbnails==2.6
plotly==5.6.0
chart-studio==1.0.0
cufflinks==0.17.3
psutil==5.8.0
requests==2.25.1
kaleido==0.2.1
pyyaml==5.4.1